Break greyhound exchanges into hpp/cpp instead of all header.
