+++ plugins/greyhound/io/Exchanges.hpp
@@ -37,7 +37,6 @@
 #include <string>
 
 #include <pdal/Dimension.hpp>
-#include <pdal/pdal_error.hpp>
 
 #include "WebSocketClient.hpp"
 #include "GreyhoundReader.hpp"
@@ -50,68 +49,19 @@ namespace exchanges
 class Exchange : public WebSocketExchange
 {
 protected:
-    Exchange(const std::string& type)
-    {
-        m_req["command"] = type;
-    }
+    Exchange(const std::string& command);
 
-    bool checkStatus() const
-    {
-        bool valid(false);
-
-        if (res().size())
-        {
-            Json::Value jsonResponse;
-            Json::Reader jsonReader;
-            jsonReader.parse(res().at(0)->get_payload(), jsonResponse);
-
-            if (jsonResponse.isMember("status") &&
-                jsonResponse["status"].isIntegral() &&
-                jsonResponse["status"].asInt() == 1)
-            {
-                valid = true;
-            }
-        }
-
-        return valid;
-    }
+    bool checkStatus() const;
 };
 
 class CreateSession : public Exchange
 {
 public:
-    CreateSession(const std::string& pipelineId)
-        : Exchange("create")
-        , m_session()
-    {
-        m_req["pipelineId"] = pipelineId;
-    }
+    CreateSession(const std::string& pipelineId);
 
-    virtual bool check()
-    {
-        bool valid(false);
+    virtual bool check();
 
-        if (checkStatus() && res().size() == 1)
-        {
-            Json::Value jsonResponse;
-            Json::Reader jsonReader;
-            jsonReader.parse(res().at(0)->get_payload(), jsonResponse);
-
-            if (jsonResponse.isMember("session") &&
-                jsonResponse["session"].isString())
-            {
-                m_session = jsonResponse["session"].asString();
-                valid = true;
-            }
-        }
-
-        return valid;
-    }
-
-    std::string getSession() const
-    {
-        return m_session;
-    }
+    std::string getSession() const;
 
 private:
     std::string m_session;
@@ -120,38 +70,11 @@ private:
 class GetNumPoints : public Exchange
 {
 public:
-    GetNumPoints(const std::string& sessionId)
-        : Exchange("pointsCount")
-        , m_count()
-    {
-        m_req["session"] = sessionId;
-    }
+    GetNumPoints(const std::string& sessionId);
 
-    virtual bool check()
-    {
-        bool valid(false);
+    virtual bool check();
 
-        if (checkStatus() && res().size() == 1)
-        {
-            Json::Value jsonResponse;
-            Json::Reader jsonReader;
-            jsonReader.parse(res().at(0)->get_payload(), jsonResponse);
-
-            if (jsonResponse.isMember("count") &&
-                jsonResponse["count"].isIntegral())
-            {
-                m_count = jsonResponse["count"].asUInt();
-                valid = true;
-            }
-        }
-
-        return valid;
-    }
-
-    std::size_t count() const
-    {
-        return m_count;
-    }
+    std::size_t count() const;
 
 private:
     std::size_t m_count;
@@ -160,63 +83,11 @@ private:
 class GetSchema : public Exchange
 {
 public:
-    GetSchema(const std::string& sessionId)
-        : Exchange("schema")
-        , m_dimData()
-    {
-        m_req["session"] = sessionId;
-    }
+    GetSchema(const std::string& sessionId);
 
-    virtual bool check()
-    {
-        bool valid(false);
+    virtual bool check();
 
-        if (checkStatus() && res().size() == 1)
-        {
-            Json::Value jsonResponse;
-            Json::Reader jsonReader;
-            jsonReader.parse(res().at(0)->get_payload(), jsonResponse);
-
-            if (jsonResponse.isMember("schema") &&
-                jsonResponse["schema"].isString())
-            {
-                Json::Value jsonSchema;
-                jsonReader.parse(jsonResponse["schema"].asString(), jsonSchema);
-
-                if (jsonSchema.isMember("dimensions") &&
-                    jsonSchema["dimensions"].isArray())
-                {
-                    Json::Value jsonDimArray(jsonSchema["dimensions"]);
-
-                    for (std::size_t i(0); i < jsonDimArray.size(); ++i)
-                    {
-                        const Json::Value& jsonDim(
-                                jsonDimArray[static_cast<Json::ArrayIndex>(i)]);
-
-                        const Dimension::Id::Enum id(
-                                Dimension::id(jsonDim["name"].asString()));
-
-                        const Dimension::Type::Enum type(
-                            static_cast<Dimension::Type::Enum>(
-                                static_cast<int>(Dimension::fromName(
-                                    jsonDim["type"].asString())) |
-                                std::stoi(jsonDim["size"].asString())));
-
-                        m_dimData.push_back(DimData(id, type));
-                    }
-
-                    valid = true;
-                }
-            }
-        }
-
-        return valid;
-    }
-
-    std::vector<DimData> schema() const
-    {
-        return m_dimData;
-    }
+    std::vector<DimData> schema() const;
 
 private:
     std::vector<DimData> m_dimData;
@@ -225,88 +96,14 @@ private:
 class Read : public Exchange
 {
 public:
-    Read(const std::string& sessionId, int offset = 0, int count = -1)
-        : Exchange("read")
-        , m_initialized(false)
-        , m_error(false)
-        , m_pointsToRead(0)
-        , m_numBytes(0)
-        , m_numBytesReceived(0)
-        , m_data()
-    {
-        m_req["session"] = sessionId;
-        m_req["start"] = offset;
-        if (count != -1) m_req["count"] = count;
-    }
+    Read(const std::string& sessionId, int offset = 0, int count = -1);
 
-    virtual bool done() const
-    {
-        return (m_initialized && m_numBytesReceived >= m_numBytes) || m_error;
-    }
+    virtual bool done() const;
+    virtual bool check();
+    virtual void handleRx(const message_ptr message);
 
-    virtual bool check()
-    {
-        bool valid(false);
-
-        if (!m_error && checkStatus() && res().size() >= 1)
-        {
-            Json::Value jsonResponse;
-            Json::Reader jsonReader;
-            jsonReader.parse(res().at(0)->get_payload(), jsonResponse);
-
-            if (jsonResponse.isMember("numPoints") &&
-                jsonResponse["numPoints"].isIntegral() &&
-                jsonResponse.isMember("numBytes") &&
-                jsonResponse["numBytes"].isIntegral())
-            {
-                m_pointsToRead =
-                    std::max<int>(jsonResponse["numPoints"].asInt(), 0);
-                m_numBytes =
-                    std::max<int>(jsonResponse["numBytes"].asInt(), 0);
-
-                valid = true;
-            }
-        }
-
-        return valid;
-    }
-
-    virtual void handleRx(const message_ptr message)
-    {
-        if (!m_initialized)
-        {
-            if (check())
-            {
-                m_initialized = true;
-            }
-            else
-            {
-                m_error = true;
-            }
-        }
-        else
-        {
-            if (message->get_opcode() == websocketpp::frame::opcode::binary)
-            {
-                m_data.push_back(&message->get_payload());
-                m_numBytesReceived += message->get_payload().length();
-            }
-            else
-            {
-                m_error = true;
-            }
-        }
-    }
-
-    std::vector<const std::string*> data() const
-    {
-        return m_data;
-    }
-
-    std::size_t numRead() const
-    {
-        return m_pointsToRead;
-    }
+    std::vector<const std::string*> data() const;
+    std::size_t numRead() const;
 
 private:
     bool m_initialized;
@@ -320,16 +117,9 @@ private:
 class Destroy: public Exchange
 {
 public:
-    Destroy(const std::string& sessionId)
-        : Exchange("destroy")
-    {
-        m_req["session"] = sessionId;
-    }
+    Destroy(const std::string& sessionId);
 
-    virtual bool check()
-    {
-        return checkStatus() && res().size() == 1;
-    }
+    virtual bool check();
 };
 
 } // namespace exchanges
